+++ contracts/interfaces/IAdapter.sol
@@ -5,9 +5,9 @@ pragma solidity 0.8.19;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IAdapter {
-    function feePromille() external view returns (uint256);
+    function feePromille() external view returns (uint8);
 
-    function setFeePromille(uint256 _feePromille) external;
+    function setFeePromille(uint8 _feePromille) external;
 
     function swap(
         address payable recipient,
